requirements.txt
python>=3.7
pygame-ce
